functions/fn_stg/requirements.txt
boto==2.49.0
boto3==1.12.49
botocore==1.15.49
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
cloudstorage==0.10.1
docutils==0.15.2
google-api-core==1.16.0
google-auth==1.13.1
google-auth-oauthlib==0.4.1
google-cloud==0.34.0
google-cloud-bigquery==1.24.0
google-cloud-core==1.3.0
google-cloud-logging==1.15.0
google-cloud-storage==1.27.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.28.1
idna==2.9
inflection==0.4.0
jmespath==0.9.5
numpy==1.18.3
oauthlib==3.1.0
pandas==1.0.3
pandas-gbq==0.13.1
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydata-google-auth==1.1.0
python-dateutil==2.8.1
python-magic==0.4.15
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
simplejson==3.17.0
six==1.14.0
smartopen==0.2
urllib3==1.25.8
smart-open==2.0.0
